% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertNumericDate}
\alias{convertNumericDate}
\title{Converts given year, month and day to UNIX time-code for API}
\usage{
convertNumericDate(YYYY, MM, DD, tz = "America/New_York")
}
\arguments{
\item{YYYY}{Year; must be 4 digits}

\item{MM}{Month}

\item{DD}{Day}

\item{tz}{Specify timezone based on R base time zones. Default: America/New_York}
}
\value{
UNIX code time-stamp
}
\description{
Converts given year, month and day to UNIX time-code for API
}
