% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividends.R
\name{yfinDividends}
\alias{yfinDividends}
\title{Title}
\usage{
yfinDividends(security, region = "US", start_date = NULL, end_date = NULL)
}
\arguments{
\item{security}{}

\item{region}{}

\item{start_date}{}

\item{end_date}{}
}
\value{

}
\description{
Title
}
