% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{yfinHistory}
\alias{yfinHistory}
\title{Gather the historical data for a given ticker}
\usage{
yfinHistory(
  security,
  region = "US",
  start_date,
  end_date,
  interval = "1d",
  historyType = "price"
)
}
\arguments{
\item{security}{The ticker for the security}

\item{region}{Default: US; selects the stock exchanges to get security from.}

\item{start_date}{Must be numeric in this format: YYYYMMDD}

\item{end_date}{Must be numeric in this format: YYYYMMDD}

\item{interval}{}

\item{type}{}
}
\value{

}
\description{
Gather the historical data for a given ticker
}
