% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profit.R
\name{yfinProfit}
\alias{yfinProfit}
\title{Title}
\usage{
yfinProfit(
  security,
  region,
  basePrice,
  quantity = 1,
  tradeDate,
  realtime = FALSE,
  futureDate = NULL,
  summary = FALSE
)
}
\arguments{
\item{security}{}

\item{region}{}

\item{basePrice}{}

\item{quantity}{}

\item{tradeDate}{}

\item{realtime}{}

\item{futureDate}{}

\item{summary}{}
}
\value{

}
\description{
Title
}
