% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{yfinQuery}
\alias{yfinQuery}
\title{Run a query against the API. This function is used as the base query for all
other functions in this package, or you can input your own custom path}
\usage{
yfinQuery(path)
}
\arguments{
\item{path}{The path of the query.}
}
\value{
Result of the query
}
\description{
Run a query against the API. This function is used as the base query for all
other functions in this package, or you can input your own custom path
}
