% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{yfinQuotes}
\alias{yfinQuotes}
\title{Pull the most recent quote for a security/ies and specify the fields to pull}
\usage{
yfinQuotes(
  securities,
  regions = "US",
  fields = c("currency", "symbol", "longName", "regularMarketPrice", "bid", "ask",
    "marketState")
)
}
\arguments{
\item{securities}{String or vector of strings, maximum 10 securities. Must be all from the same region.}

\item{fields}{Default c("currency", "symbol", "longName", "regularMarketPrice", "bid", "ask", "marketState")}

\item{region}{Default: US; selects the stock exchanges to get security from.}
}
\value{
Data frame of stock market quote data.
}
\description{
Pull the most recent quote for a security/ies and specify the fields to pull
}
