% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{yfinTestConnection}
\alias{yfinTestConnection}
\title{This function tests the API connection.}
\usage{
yfinTestConnection()
}
\value{
Returns the status code of the test API call.
}
\description{
This function tests the API connection.
}
